refactor: ♻️ move workspace on its own class
